# cmd/testserver/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Optional HTTP harness for the storagepool Go library. Authenticate with Basic Auth to get a JWT, then call pool APIs with Bearer token and Hypervisor-Uri header. Use Authorize in Swagger UI to set credentials (Postman-like try-it-out).",
        "title": "ViRest Storage Pool Testserver",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/storage-pool/authenticate": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Authenticate with HTTP Basic and receive a Bearer JWT",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Issue JWT",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.TokenEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.Envelope"
                        }
                    }
                }
            }
        },
        "/storage-pool/autostart": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pools"
                ],
                "summary": "Set pool autostart",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Libvirt URI",
                        "name": "Hypervisor-Uri",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Autostart settings",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AutostartRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.Envelope"
                        }
                    }
                }
            }
        },
        "/storage-pool/build": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pools"
                ],
                "summary": "Build pool storage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Libvirt URI",
                        "name": "Hypervisor-Uri",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Pool UUID and build flags",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UuidOptionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.Envelope"
                        }
                    }
                }
            }
        },
        "/storage-pool/capabilities": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pools"
                ],
                "summary": "Get storage pool capabilities",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Libvirt URI",
                        "name": "Hypervisor-Uri",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.Envelope"
                        }
                    }
                }
            }
        },
        "/storage-pool/create": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pools"
                ],
                "summary": "Create (start) a pool",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Libvirt URI",
                        "name": "Hypervisor-Uri",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Pool UUID and create flags",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UuidOptionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.Envelope"
                        }
                    }
                }
            }
        },
        "/storage-pool/define": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pools"
                ],
                "summary": "Define a storage pool",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Libvirt URI",
                        "name": "Hypervisor-Uri",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Pool definition",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.DefineRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.Envelope"
                        }
                    }
                }
            }
        },
        "/storage-pool/delete": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pools"
                ],
                "summary": "Delete pool resources",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Libvirt URI",
                        "name": "Hypervisor-Uri",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Pool UUID and delete flags",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UuidOptionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.Envelope"
                        }
                    }
                }
            }
        },
        "/storage-pool/destroy": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pools"
                ],
                "summary": "Destroy (stop) a pool",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Libvirt URI",
                        "name": "Hypervisor-Uri",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Pool UUID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UuidOptionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.Envelope"
                        }
                    }
                }
            }
        },
        "/storage-pool/detail": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pools"
                ],
                "summary": "Get pool detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Libvirt URI",
                        "name": "Hypervisor-Uri",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pool UUID",
                        "name": "uuid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "XML flags",
                        "name": "option",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.Envelope"
                        }
                    }
                }
            }
        },
...
        "/storage-pool/get-gid": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meta"
                ],
                "summary": "Process GID of the testserver",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Libvirt URI (allowlist check)",
                        "name": "Hypervisor-Uri",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.Envelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.DefineRequest": {
            "type": "object",
            "properties": {
                "option": {
                    "type": "integer",
                    "example": 0
                },
                "storagePool": {}
            }
        },
        "handlers.Envelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {
                    "type": "string"
                },
                "response": {
                    "type": "boolean"
                }
            }
        },
...
        "handlers.TokenEnvelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object",
                    "properties": {
                        "token": {
                            "type": "string"
                        }
                    }
                },
                "error": {
                    "type": "string"
                },
                "response": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "description": "JWT from /storage-pool/authenticate. Prefer value: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}